Check working folder before application and service commands
