+++ src/commands/create-deployment-profile.ts
@@ -3,11 +3,18 @@ import * as path from "path";
 import * as constants from "../constants";
 import * as fileUtils from '../utils/file';
 import DeploymentProfile from '../utils/deployment-profile';
+import { window } from 'vscode';
+import { getWorkingFolder } from '../utils/workspace';
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 
 export async function createDeploymentProfile(templateFile: string, templateUri: string) {
+	const cwd = await getWorkingFolder();
+	if (!cwd) {
+		window.showErrorMessage('Please open a workspace directory first.');
+		return;
+    }
     var configuration = vscode.workspace.getConfiguration(constants.config.projectName);
     let deployment = new DeploymentProfile();
     deployment.SubscriptionId = configuration.get<string>(constants.config.defaultSubscriptionId);
